rename contracts package
