+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/LoanApplicationRoot.java
@@ -1,8 +1,8 @@
 package com.asc.loanservice.domain.loan.application;
 
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationCreateInput;
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationEvaluationStatus;
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationResult;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationCreateInput;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationEvaluationStatus;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationResult;
 import com.asc.loanservice.infrastructure.LoanEvaluationResult;
 import com.asc.loanservice.infrastructure.repository.model.LoanApplication;
 import lombok.Builder;
