+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/application/port/LoanEvaluatorProviderPort.java
@@ -1,6 +1,6 @@
 package com.asc.loanservice.domain.loan.application.port;
 
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationRequest;
 import com.asc.loanservice.infrastructure.LoanEvaluationResult;
 
 public interface LoanEvaluatorProviderPort {
