+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/rule/MaximumInstallmentAmountRule.java
@@ -1,6 +1,6 @@
 package com.asc.loanservice.domain.loan.rule;
 
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationRequest;
 
 public class MaximumInstallmentAmountRule implements Rule {
