+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/rule/NotExistInDeptRegisterRule.java
@@ -3,7 +3,7 @@ package com.asc.loanservice.domain.loan.rule;
 
 import com.asc.loanservice.domain.loan.rule.port.DebtPort;
 import com.asc.loanservice.external.DebtCustomerResult;
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationRequest;
 import lombok.AllArgsConstructor;
