+++ src/loan-service/src/main/java/com/asc/loanservice/domain/loan/rule/Rule.java
@@ -1,6 +1,6 @@
 package com.asc.loanservice.domain.loan.rule;
 
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationRequest;
 
 public interface Rule {
     Boolean isValid(LoanApplicationRequest loanApplicationRequest);
