+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/api/LoanApplicationController.java
@@ -1,8 +1,8 @@
 package com.asc.loanservice.infrastructure.api;
 
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationView;
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationResult;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationView;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationResult;
 import com.asc.loanservice.domain.loan.application.LoanApplicationFacade;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
