+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/api/contract/LoanApplicationCreateInput.java
@@ -1,17 +1,19 @@
-package com.asc.loanservice.infrastructure.api.contracts;
+package com.asc.loanservice.infrastructure.api.contract;
 
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Getter
-@Builder
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class LoanApplicationView {
+public class LoanApplicationCreateInput {
     private String loanRequestNumber;
     private String customerName;
     private LocalDate customerBirthday;
@@ -20,6 +22,4 @@ public class LoanApplicationView {
     private BigDecimal loanAmount;
     private Integer numberOfInstallments;
     private LocalDate firstInstallmentDate;
-    private LoanApplicationEvaluationStatus evaluationResult;
-    private LocalDateTime registrationDate;
 }
