+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/api/contract/LoanApplicationEvaluationStatus.java
@@ -1,4 +1,4 @@
-package com.asc.loanservice.infrastructure.api.contracts;
+package com.asc.loanservice.infrastructure.api.contract;
 
 public enum LoanApplicationEvaluationStatus {
     APPROVED,
