+++ src/loan-service/src/main/java/com/asc/loanservice/infrastructure/mapper/LoanMapperAdapter.java
@@ -1,7 +1,7 @@
 package com.asc.loanservice.infrastructure.mapper;
 
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationCreateInput;
-import com.asc.loanservice.infrastructure.api.contracts.LoanApplicationRequest;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationCreateInput;
+import com.asc.loanservice.infrastructure.api.contract.LoanApplicationRequest;
 import com.asc.loanservice.domain.loan.application.port.LoanMapperPort;
 import org.mapstruct.Mapper;
 import org.springframework.stereotype.Component;
